# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehaus API",
        "description": "REST API v1 untuk sistem manajemen inventaris Warehaus. Setiap endpoint mengembalikan envelope { success, message, data }.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Warehaus API server"
        }
    ],
    "paths": {
        "/api/v1/borrowings": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "List borrowings",
                "operationId": "4da3fd4c62369880bdfc1e1dd8569b5b",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dipinjam",
                                "dikembalikan",
                                "terlambat"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Create a borrowing (transactional; stock is decremented atomically)",
                "operationId": "d456124abb9f2cdd9cc4564cfb3d9f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "borrower_name",
                                    "borrowed_at",
                                    "due_at",
                                    "items"
                                ],
                                "properties": {
                                    "borrower_name": {
                                        "type": "string",
                                        "example": "Budi"
                                    },
                                    "borrowed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-05"
                                    },
                                    "due_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-12"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2,
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Insufficient stock or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/borrowings/{borrowing}": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Get a single borrowing",
                "operationId": "2aa7efb73c8df64e9ceb8db6ad9cc926",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Update notes or due date",
                "operationId": "9acbfd352e34e9158bc315902ddc032e",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "due_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Delete a returned borrowing",
                "operationId": "2fb578fd9b0614773b34c003036f84bc",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Borrowing still active",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/borrowings/{borrowing}/return": {
            "patch": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Mark items as returned and restock (transactional)",
                "operationId": "abd6ddddc892bc173c99fd8ad2bf64fa",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "description": "Base OpenAPI info + security scheme for the Warehaus API.\n\nKept as an empty controller purely so the l5-swagger annotation scanner\nhas a well-known place to pick up the top-level metadata.",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validasi gagal."
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login via POST /api/v1/login to obtain a token.",
                "bearerFormat": "Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}